Revert "Add a missing default mapping (cX -> cx$)"
This reverts commit 3968b301cf93ec3efd2b251fea60ebeabbd3171c.

There's a reason not to include this by default (in this way):
If someone wants to rebind the default cx, cxx and {visual}X, the cX
will remain but in a broken way.

Also I'm adding another vim-exchange map in my configs that's more odd,
so I'll move cX to my keymaps too.